Add PR logic that works for a sample moviedb repo fork
